pyproject.toml
[tool.poetry]
name = "mptreasury"
version = "0.1.0"
description = ""
authors = ["crpier <crpier42@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
pydantic = "^1.9.0"
alembic = "^1.7.7"
sqlalchemy2-stubs = "^0.0.2-alpha.22"
python3-discogs-client = "^2.3.14"
fuzzywuzzy = "^0.18.0"
python-Levenshtein = "^0.12.2"
mutagen = "^1.45.1"

[tool.poetry.dev-dependencies]
pytest = "^7.1.1"
pytest-memray = "^1.0.0"
memray = "^1.0.3"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.pytest.ini_options]
log_cli = true
log_cli_level = "INFO"
markers = [
  "unit: self explanatory I think",
  "component: self explanatory I think",
  "e2e: self explanatory I think",
]

